8.1.4.1 Fixed bugs. Completed updating with server.
